+++ nixie_firmware/src/tube_driver.hpp
@@ -17,14 +17,14 @@
 class TubeDriver
 {
 private:
-    int brightness;
-    int l_dot_brightness;
-    int r_dot_brightness;
+    int16_t brightness;
+    int16_t l_dot_brightness;
+    int16_t r_dot_brightness;
     bool status;
     EveryTimer cathode_poisoning_cycle;
     SensorDriver *sensor_driver;
 
-    void set_tube_brightness(int, int, int);
+    void set_tube_brightness(int16_t, int16_t, int16_t);
 
 public:
     TubeDriver(SensorDriver *);
@@ -33,13 +33,13 @@ public:
 
     void loop();
 
-    void set_tubes(int, int, int, int, int, int);
+    void set_tubes(int8_t, int8_t, int8_t, int8_t, int8_t, int8_t);
 
-    void set_brightness(int);
+    void set_brightness(int16_t);
 
-    void set_dots_brightness(int, int);
+    void set_dots_brightness(int16_t, int16_t);
 
-    void display_time_and_date(int, int, int, bool);
+    void display_time_and_date(int8_t, int8_t, int8_t, bool);
 
     void display_temperature(float);
 
@@ -47,7 +47,7 @@ public:
 
     void turn_off(bool);
 
-    void turn_on(int);
+    void turn_on(int16_t);
 
     void cathode_poisoning_prevention(unsigned long);
 };
